Added new group for MembershipTeam, this group can approve membership etc but not see address or contact details. Fixed various other permission issues.

+++ app/View/Members/edit.ctp
@@ -18,13 +18,21 @@
 	echo $this->Form->input('email');
 	echo $this->Form->input('unlock_text');
 
+	if(array_key_exists('address', $member))
+	{
 
-	echo $this->Form->input('address_1', array( 'label' => 'Address part 1 (House name/number and street)' ) );
-	echo $this->Form->input('address_2', array( 'label' => 'Address part 2' ) );
-	echo $this->Form->input('address_city', array( 'label' => 'City' ) );
-	echo $this->Form->input('address_postcode', array( 'label' => 'Postcode' ) );
+		echo $this->Form->input('address_1', array( 'label' => 'Address part 1 (House name/number and street)' ) );
+		echo $this->Form->input('address_2', array( 'label' => 'Address part 2' ) );
+		echo $this->Form->input('address_city', array( 'label' => 'City' ) );
+		echo $this->Form->input('address_postcode', array( 'label' => 'Postcode' ) );
 
-	echo $this->Form->input('contact_number' );
+	}
+
+	if(array_key_exists('contactNumber', $member))
+	{
+		echo $this->Form->input('contact_number' );
+	}
+
 
 	if( isset($member['paymentRef']) )
 	{
